lint fix

+++ contracts/avalanche/connectors/aave/v2/main.sol
@@ -267,7 +267,7 @@ abstract contract AaveResolver is Events, Helpers {
 
 		for (uint256 i = 0; i < _length; i++) {
 			bool isAvax = tokens[i] == avaxAddr;
-		    address _token = isAvax ? wavaxAddr : tokens[i];
+			address _token = isAvax ? wavaxAddr : tokens[i];
 
 			if (getCollateralBalance(_token) > 0 && !getIsColl(_token)) {
 				aave.setUserUseReserveAsCollateral(_token, true);
@@ -278,28 +278,29 @@ abstract contract AaveResolver is Events, Helpers {
 		_eventParam = abi.encode(tokens);
 	}
 
-    /**
-     * @dev Swap borrow rate mode
-     * @notice Swaps user borrow rate mode between variable and stable
-     * @param token The address of the token to swap borrow rate.(For AVAX: 0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE)
-     * @param currentRateMode Current Rate mode. (Stable = 1, Variable = 2)
-    */
-    function swapBorrowRateMode(
-        address token,
-        uint currentRateMode
-    ) external payable returns (string memory _eventName, bytes memory _eventParam) {
-        AaveInterface aave = AaveInterface(aaveProvider.getLendingPool());
+	/**
+	 * @dev Swap borrow rate mode
+	 * @notice Swaps user borrow rate mode between variable and stable
+	 * @param token The address of the token to swap borrow rate.(For AVAX: 0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE)
+	 * @param currentRateMode Current Rate mode. (Stable = 1, Variable = 2)
+	 */
+	function swapBorrowRateMode(address token, uint256 currentRateMode)
+		external
+		payable
+		returns (string memory _eventName, bytes memory _eventParam)
+	{
+		AaveInterface aave = AaveInterface(aaveProvider.getLendingPool());
 
 		bool isAVAX = token == avaxAddr;
 		address _token = isAVAX ? wavaxAddr : token;
 
-        if (getPaybackBalance(_token, currentRateMode) > 0) {
-            aave.swapBorrowRateMode(_token, currentRateMode);
-        }
+		if (getPaybackBalance(_token, currentRateMode) > 0) {
+			aave.swapBorrowRateMode(_token, currentRateMode);
+		}
 
-        _eventName = "LogSwapRateMode(address,uint256)";
-        _eventParam = abi.encode(token, currentRateMode);
-    }
+		_eventName = "LogSwapRateMode(address,uint256)";
+		_eventParam = abi.encode(token, currentRateMode);
+	}
 }
 
 contract ConnectV2AaveV2Avalanche is AaveResolver {
